refactor(ghostmap): split suppression finder/Add helpers into GhostMapPresence.SuppressionFinders.cs
